Seems to shut down cleanly now
git-svn-id: https://svn.ee.ucl.ac.uk/repos/jefferson/userspacerouter@180 83772449-cf8d-4b79-a1b1-a7e526131002

+++ usr/localcontroller/LocalControllerInfo.java
@@ -40,7 +40,7 @@ public class LocalControllerInfo extends LocalHostInfo {
     }
     
     public double getUsage() {
-        double usage= currRouters_/maxRouters_;
+        double usage= (double)currRouters_/maxRouters_;
         return usage;
     }
